# core/aggregator/api/swagger.yaml
openapi: 3.0.0
info:
  title: Swagger Gandalf
  description: "This is a sample Petstore server.  You can find \nout more about Swagger\
    \ at \n[http://swagger.io](http://swagger.io) or on \n[irc.freenode.net, #swagger](http://swagger.io/irc/).\n"
  termsOfService: http://swagger.io/terms/
  contact:
    email: romain.fairant@orness.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0-oas3
...
servers:
- url: https://virtserver.swaggerhub.com/ditrit/Gandalf/1.0.0
- url: http://virtserver.swaggerhub.com/ditrit/Gandalf/1.0.0
...
paths:
...
  /resource:
    get:
      tags:
      - resource
      summary: List resources
      operationId: listResource
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/inline_response_200_4'
                x-content-type: application/json
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    post:
      tags:
      - resource
      summary: Create a resource
      operationId: createResource
      requestBody:
        $ref: '#/components/requestBodies/createOrUpdateResource'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        "400":
          description: Invalid request payload
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
  /resource/{resourceId}:
    get:
      tags:
      - resource
      summary: Find resource by ID
      description: Returns a single resource
      operationId: getResourceById
      parameters:
      - name: resourceId
        in: path
        description: ID of role to return
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        "400":
          description: Invalid ID supplied
        "404":
          description: Resource not found
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    put:
      tags:
      - resource
      summary: Updates a resource
      operationId: updateResource
      parameters:
      - name: resourceId
        in: path
        description: ID of resource that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        $ref: '#/components/requestBodies/createOrUpdateResource'
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        "400":
          description: Invalid ID supplied | Invalid request payload
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    delete:
      tags:
      - resource
      summary: Deletes a resource
      operationId: deleteResource
      parameters:
      - name: api_key
        in: header
        required: false
        style: simple
        explode: false
        schema:
          type: string
      - name: resourceId
        in: path
        description: Resource id to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Success
        "400":
          description: Invalid ID supplied
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
  /resource/{resourceName}:
    get:
      tags:
      - resource
      summary: Find resource by Name
      description: Returns a single Resource
      operationId: getResourceByName
      parameters:
      - name: resourceName
        in: path
        description: Name of resource to return
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        "404":
          description: Resource not found
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
  /resourceType:
    get:
      tags:
      - resourceType
      summary: List resourceTypes
      operationId: listResourceType
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/inline_response_200_5'
                x-content-type: application/json
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    post:
      tags:
      - resourceType
      summary: Create a resourcetype
      operationId: createResourceType
      requestBody:
        $ref: '#/components/requestBodies/createOrUpdateResourceType'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceType'
        "400":
          description: Invalid request payload
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
  /resourceType/{resourceTypeId}:
    get:
      tags:
      - resourceType
      summary: Find resourceType by ID
      description: Returns a single resourceType
      operationId: getResourceTypeById
      parameters:
      - name: resourceTypeId
        in: path
        description: ID of role to return
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceType'
        "400":
          description: Invalid ID supplied
        "404":
          description: ResourceType not found
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    put:
      tags:
      - resourceType
      summary: Updates a resourcetype
      operationId: updateResourceType
      parameters:
      - name: resourceTypeId
        in: path
        description: ID of resourcetype that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        $ref: '#/components/requestBodies/createOrUpdateResourceType'
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceType'
        "400":
          description: Invalid ID supplied | Invalid request payload
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    delete:
      tags:
      - resourceType
      summary: Deletes a resourceType
      operationId: deleteResourceType
      parameters:
      - name: api_key
        in: header
        required: false
        style: simple
        explode: false
        schema:
          type: string
      - name: resourceTypeId
        in: path
        description: ResourceType id to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Success
        "400":
          description: Invalid ID supplied
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
  /resourceType/{resourceTypeName}:
    get:
      tags:
      - resourceType
      summary: Find resourceType by Name
      description: Returns a single ResourceType
      operationId: getResourceTypeByName
      parameters:
      - name: resourceTypeName
        in: path
        description: Name of resourceType to return
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceType'
        "404":
          description: ResourceType not found
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
...
components:
  schemas:
    Model:
      type: object
      properties:
        id:
          type: integer
          description: The ID.
        createdAt:
          type: string
          description: The created date.
          format: date-time
        updatedAt:
          type: string
          description: The updated date.
          format: date-time
        deletedAt:
          type: string
          description: The deleted date.
          format: date-time
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        deletedAt: 2000-01-23T04:56:07.000+00:00
        id: 0
        updatedAt: 2000-01-23T04:56:07.000+00:00
...
    Resource:
      type: object
      properties:
        model:
          $ref: '#/components/schemas/Model'
        name:
          type: string
          description: The resource name.
        logicalComponentId:
          type: integer
          description: The logical component id.
        domainId:
          type: integer
          description: The domain id.
        resourceTypeId:
          type: integer
          description: The resourceType id.
      example:
        resourceTypeId: 7
        name: name
        model:
          createdAt: 2000-01-23T04:56:07.000+00:00
          deletedAt: 2000-01-23T04:56:07.000+00:00
          id: 0
          updatedAt: 2000-01-23T04:56:07.000+00:00
        logicalComponentId: 5
        domainId: 2
    ResourceType:
      type: object
      properties:
        model:
          $ref: '#/components/schemas/Model'
        name:
          type: string
          description: The resourceType name.
        pivotId:
          type: integer
          description: The pivot id.
        productConnectorId:
          type: integer
          description: The productConnector id.
      example:
        productConnectorId: 6
        pivotId: 0
        name: name
        model:
          createdAt: 2000-01-23T04:56:07.000+00:00
          deletedAt: 2000-01-23T04:56:07.000+00:00
          id: 0
          updatedAt: 2000-01-23T04:56:07.000+00:00
...
    inline_response_200_4:
      type: object
      properties:
        schema:
          $ref: '#/components/schemas/Resource'
      example:
        schema:
          resourceTypeId: 7
          name: name
          model:
            createdAt: 2000-01-23T04:56:07.000+00:00
            deletedAt: 2000-01-23T04:56:07.000+00:00
            id: 0
            updatedAt: 2000-01-23T04:56:07.000+00:00
          logicalComponentId: 5
          domainId: 2
    inline_response_200_5:
      type: object
      properties:
        schema:
          $ref: '#/components/schemas/ResourceType'
      example:
        schema:
          productConnectorId: 6
          pivotId: 0
          name: name
          model:
            createdAt: 2000-01-23T04:56:07.000+00:00
            deletedAt: 2000-01-23T04:56:07.000+00:00
            id: 0
            updatedAt: 2000-01-23T04:56:07.000+00:00
...
  requestBodies:
...
    createOrUpdateResource:
      description: The resource to create.
      content:
        application/json:
          schema:
            required:
            - domainID
            - logicalComponentID
            - name
            - resourceTypeID
            type: object
            properties:
              name:
                type: string
              logicalComponentID:
                type: integer
              domainID:
                type: integer
              resourceTypeID:
                type: integer
    createOrUpdateResourceType:
      description: The resourceType to create.
      content:
        application/json:
          schema:
            required:
            - name
            - pivotID
            - productConnectorID
            type: object
            properties:
              name:
                type: string
              pivotID:
                type: integer
              productConnectorID:
                type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT